fixed mapstruct, added correct datatypes in DTOs and models

+++ src/main/java/accelerate/alumni/alumnibackend/model/dtos/post/PostDTO.java
@@ -10,16 +10,16 @@ import java.util.Set;
 @Getter
 @Setter
 public class PostDTO {
-    private int id;
+    private Long id;
     private java.time.ZonedDateTime createdAt;
     private java.time.ZonedDateTime updatedAt;
     private String title;
     private String content;
     private String postTarget;
     private SenderDTO senderId;
-    private int originId;
-    private int replyParentId;
-    private Set<Integer> replies;
+    private Long originId;
+    private Long replyParentId;
+    private Set<Long> replies;
     private SenderDTO targetUser;
     private GroupMiniDTO targetGroup;
 }
